+++ src/connectors/ie-tii/index.ts
@@ -1,0 +1,5 @@
+import { TiiConnector } from "./tii.connector.js";
+
+export function createTiiConnector() {
+  return new TiiConnector();
+}